reviews

+++ BackEnd/src/main/java/com/cars24/fullstack/data/dto/ProductDto.java
@@ -2,12 +2,16 @@ package com.cars24.fullstack.data.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductDto {
     private String id;
     private String name;
     private String image;
     private String seller;
-    private double price;
+    private BigDecimal price;
     private String description;
+
+
 }
